Major refactor moving to a single package path, and support for multiple lines in requests
