--- internal/gateways/gateways.go
@@ -1,107 +1,0 @@
-// Package gateways provides the VeriFactu gateway
-package gateways
-
-import (
-	"context"
-	"crypto/tls"
-	"crypto/x509"
-	"fmt"
-	"net/http"
-	"os"
-	"strconv"
-
-	"github.com/go-resty/resty/v2"
-	"github.com/invopop/gobl.verifactu/doc"
-	"github.com/invopop/xmldsig"
-)
-
-// Environment defines the environment to use for connections
-type Environment string
-
-// Environment to use for connections
-const (
-	EnvironmentProduction Environment = "production"
-	EnvironmentSandbox    Environment = "sandbox"
-
-	ProductionBaseURL = "https://www1.agenciatributaria.gob.es/wlpl/TIKE-CONT/ws/SistemaFacturacion/VerifactuSOAP"
-	TestingBaseURL    = "https://prewww1.aeat.es/wlpl/TIKE-CONT/ws/SistemaFacturacion/VerifactuSOAP"
-
-	statusCorrect            = "Correcto"
-	statusAcceptedWithErrors = "AceptadoConErrores"
-	statusCancelled          = "Anulada"
-	statusIncorrect          = "Incorrecto"
-)
-
-// Connection defines what is expected from a connection to a gateway.
-type Connection struct {
-	client *resty.Client
-}
-
-// New instantiates and configures a new connection to the VeriFactu gateway.
-func New(env Environment, cert *xmldsig.Certificate) (*Connection, error) {
-	// Prepare the tls configuration
-	tlsConf, err := cert.TLSAuthConfig()
-	if err != nil {
-		return nil, doc.ErrValidation.WithMessage(fmt.Errorf("preparing TLS config: %v", err).Error())
-	}
-	certs, err := x509.SystemCertPool()
-	if err != nil {
-		return nil, fmt.Errorf("preparing cert pool: %w", err)
-	}
-	tlsConf.RootCAs = certs
-	tlsConf.Renegotiation = tls.RenegotiateOnceAsClient
-
-	c := new(Connection)
-	c.client = resty.New()
-
-	switch env {
-	case EnvironmentProduction:
-		c.client.SetBaseURL(ProductionBaseURL)
-	default:
-		c.client.SetBaseURL(TestingBaseURL)
-	}
-	// tlsConf.InsecureSkipVerify = true
-	c.client.SetTLSClientConfig(tlsConf)
-	c.client.SetDebug(os.Getenv("DEBUG") == "true")
-	return c, nil
-}
-
-// Post sends the VeriFactu document to the gateway
-func (c *Connection) Post(ctx context.Context, pyl []byte) error {
-	return c.post(ctx, TestingBaseURL, pyl)
-}
-
-func (c *Connection) post(ctx context.Context, path string, payload []byte) error {
-	out := new(Envelope)
-	req := c.client.R().
-		SetContext(ctx).
-		SetHeader("Content-Type", "application/xml").
-		SetContentLength(true).
-		SetBody(payload).
-		SetResult(out)
-
-	res, err := req.Post(path)
-	if err != nil {
-		return err
-	}
-	if res.StatusCode() != http.StatusOK {
-		return doc.ErrValidation.WithCode(strconv.Itoa(res.StatusCode())).WithMessage(res.String())
-	}
-	if out.Body.Fault != nil {
-		return doc.ErrValidation.WithMessage(out.Body.Fault.Message).WithCode(out.Body.Fault.Code)
-	}
-	if r := out.Body.Respuesta; r != nil {
-		switch r.Status() {
-		case statusCorrect, statusCancelled:
-			// All good.
-			return nil
-		case statusAcceptedWithErrors:
-			// Provide a warning message
-			return doc.ErrWarning.WithCode(r.Code()).WithMessage(r.Message())
-		default:
-			return doc.ErrValidation.WithCode(r.Code()).WithMessage(r.Message())
-		}
-	}
-
-	return nil
-}
