+++ party.go
@@ -1,4 +1,4 @@
-package doc
+package verifactu
 
 import (
 	"fmt"
